requirements.txt
# ─────────────────────────────────────────────
#  Requirements — Multilingual Hate & Misinfo Detector
# ─────────────────────────────────────────────

# Core ML
torch==2.2.2
transformers==4.40.0
datasets==2.19.0

# Data processing
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.14.0
pyyaml==6.0.1

# Explainability
lime==0.2.0.1

# Visualization
matplotlib==3.9.0
seaborn==0.13.2

# MLOps
mlflow==2.13.0
dvc==3.51.2

# Web app
streamlit==1.35.0

# Utilities
tqdm==4.66.4
